backend/requirements.txt
fastapi[standard]==0.115.0
uvicorn[standard]==0.30.6
python-dotenv==1.0.1
httpx==0.28.1
pydantic==2.10.4

# Firebase
firebase-admin==6.5.0
google-cloud-firestore==2.19.0
google-cloud-storage==2.18.2

# Auth
python-jose[cryptography]==3.3.0
authlib==1.3.2

# Stripe
stripe==10.12.0

# LangGraph / LangChain
langgraph==0.2.60
langchain-core==0.3.35
langchain-together==0.3.1
google-genai==1.68.0

# Browser automation
browser-use==0.1.40
playwright==1.49.0

# Data processing
pandas==2.2.3
